+++ deployer/deployer.go
@@ -15,14 +15,13 @@ import (
 	"github.com/pentops/o5-go/deployer/v1/deployer_pb"
 	"github.com/pentops/o5-go/deployer/v1/deployer_tpb"
 	"github.com/pentops/o5-go/environment/v1/environment_pb"
-	"google.golang.org/protobuf/types/known/emptypb"
-	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 type Environment struct {
 	Environment *environment_pb.Environment
 	AWS         *environment_pb.AWS
 }
+
 type Deployer struct {
 	RotateSecrets bool
 	CancelUpdates bool
@@ -31,8 +30,6 @@ type Deployer struct {
 	cfTemplateBucket string
 
 	storage DeployerStorage
-
-	*deployer_tpb.UnimplementedDeployerTopicServer
 }
 
 func NewDeployer(storage DeployerStorage, cfTemplateBucket string, s3Client awsinfra.S3API) (*Deployer, error) {
@@ -54,127 +51,85 @@ func (d *Deployer) BeginDeployments(ctx context.Context, app *app.BuiltApplicati
 }
 
 func (d *Deployer) BeginDeployment(ctx context.Context, app *app.BuiltApplication, envName string) error {
+	trigger, err := d.BuildTrigger(ctx, app, envName)
+	if err != nil {
+		return err
+	}
+
+	return d.storage.Transact(ctx, func(ctx context.Context, tx TransitionTransaction) error {
+		return tx.PublishEvent(ctx, trigger)
+	})
+}
+
+func (dd *Deployer) BuildTrigger(ctx context.Context, app *app.BuiltApplication, envName string) (*deployer_tpb.TriggerDeploymentMessage, error) {
 
 	ctx = log.WithFields(ctx, map[string]interface{}{
 		"appName":     app.Name,
 		"environment": envName,
 	})
 
-	environment, err := d.storage.GetEnvironment(ctx, envName)
+	environment, err := dd.storage.GetEnvironment(ctx, envName)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	awsEnv := environment.GetAws()
 	if awsEnv == nil {
-		return fmt.Errorf("environment %s is not an AWS environment", envName)
+		return nil, fmt.Errorf("environment %s is not an AWS environment", envName)
 	}
 
 	deploymentID := uuid.NewString()
 
 	templateJSON, err := app.TemplateJSON()
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	templateKey := fmt.Sprintf("%s/%s/%s.json", environment.FullName, app.Name, deploymentID)
-	_, err = d.s3Client.PutObject(ctx, &s3.PutObjectInput{
-		Bucket: aws.String(d.cfTemplateBucket),
+	_, err = dd.s3Client.PutObject(ctx, &s3.PutObjectInput{
+		Bucket: aws.String(dd.cfTemplateBucket),
 		Key:    aws.String(templateKey),
 		Body:   bytes.NewReader(templateJSON),
 	})
 
 	if err != nil {
-		return err
+		return nil, err
 	}
 
-	templateURL := fmt.Sprintf("https://s3.us-east-1.amazonaws.com/%s/%s", d.cfTemplateBucket, templateKey)
+	postgresDatabases := app.PostgresDatabases()
+
+	for _, db := range postgresDatabases {
+		pgdb := db.Database.GetPostgres()
+		for _, host := range awsEnv.RdsHosts {
+			if host.ServerGroup == pgdb.ServerGroup {
+				db.RdsHost = host
+				break
+			}
+		}
+		if db.RdsHost == nil {
+			return nil, fmt.Errorf("no RDS host found for database %s", db.Database.Name)
+		}
+
+	}
+
+	templateURL := fmt.Sprintf("https://s3.us-east-1.amazonaws.com/%s/%s", dd.cfTemplateBucket, templateKey)
 
 	spec := &deployer_pb.DeploymentSpec{
 		AppName:         app.Name,
 		Version:         app.Version,
 		EnvironmentName: environment.FullName,
 		TemplateUrl:     templateURL,
-		Databases:       app.PostgresDatabases(),
+		Databases:       postgresDatabases,
 		Parameters:      app.Parameters(),
 		SnsTopics:       app.SNSTopics,
 
-		CancelUpdates:     d.CancelUpdates,
-		RotateCredentials: d.RotateSecrets,
+		CancelUpdates:     dd.CancelUpdates,
+		RotateCredentials: dd.RotateSecrets,
+		EcsCluster:        awsEnv.EcsClusterName,
 	}
 
-	return d.storage.Transact(ctx, func(ctx context.Context, tx TransitionTransaction) error {
-		return tx.PublishEvent(ctx, &deployer_tpb.TriggerDeploymentMessage{
-			DeploymentId: deploymentID,
-			Spec:         spec,
-		})
-	})
-}
-
-func (dd *Deployer) TriggerDeployment(ctx context.Context, msg *deployer_tpb.TriggerDeploymentMessage) (*emptypb.Empty, error) {
-
-	deploymentEvent := &deployer_pb.DeploymentEvent{
-		DeploymentId: msg.DeploymentId,
-		Metadata: &deployer_pb.EventMetadata{
-			EventId:   uuid.NewString(),
-			Timestamp: timestamppb.Now(),
-		},
-		Event: &deployer_pb.DeploymentEventType{
-			Type: &deployer_pb.DeploymentEventType_Triggered_{
-				Triggered: &deployer_pb.DeploymentEventType_Triggered{
-					Spec: msg.Spec,
-				},
-			},
-		},
-	}
-
-	if err := dd.RegisterEvent(ctx, deploymentEvent); err != nil {
-		return nil, err
-	}
-
-	return &emptypb.Empty{}, nil
-}
-
-func (dd *Deployer) StackStatusChanged(ctx context.Context, msg *deployer_tpb.StackStatusChangedMessage) (*emptypb.Empty, error) {
-
-	event := &deployer_pb.DeploymentEvent{
-		DeploymentId: msg.StackId.DeploymentId,
-		Metadata: &deployer_pb.EventMetadata{
-			EventId:   uuid.NewString(),
-			Timestamp: timestamppb.Now(),
-		},
-		Event: &deployer_pb.DeploymentEventType{
-			Type: &deployer_pb.DeploymentEventType_StackStatus_{
-				StackStatus: &deployer_pb.DeploymentEventType_StackStatus{
-					Lifecycle:   msg.Lifecycle,
-					FullStatus:  msg.Status,
-					StackOutput: msg.Outputs,
-				},
-			},
-		},
-	}
-	return &emptypb.Empty{}, dd.RegisterEvent(ctx, event)
-
-}
-
-func (dd *Deployer) MigrationStatusChanged(ctx context.Context, msg *deployer_tpb.MigrationStatusChangedMessage) (*emptypb.Empty, error) {
-
-	event := &deployer_pb.DeploymentEvent{
-		DeploymentId: msg.DeploymentId,
-		Metadata: &deployer_pb.EventMetadata{
-			EventId:   uuid.NewString(),
-			Timestamp: timestamppb.Now(),
-		},
-
-		Event: &deployer_pb.DeploymentEventType{
-			Type: &deployer_pb.DeploymentEventType_DbMigrateStatus{
-				DbMigrateStatus: &deployer_pb.DeploymentEventType_DBMigrateStatus{
-					MigrationId: msg.MigrationId,
-					Status:      msg.Status,
-				},
-			},
-		},
-	}
-	return &emptypb.Empty{}, dd.RegisterEvent(ctx, event)
-
+	return &deployer_tpb.TriggerDeploymentMessage{
+		DeploymentId: deploymentID,
+		Spec:         spec,
+	}, nil
 }
