+++ localrun/eventloop.go
@@ -4,13 +4,14 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"sync"
 
 	"github.com/bufbuild/protovalidate-go"
 	"github.com/pentops/log.go/log"
 	"github.com/pentops/o5-deploy-aws/awsinfra"
+	"github.com/pentops/o5-deploy-aws/deployer"
+	"github.com/pentops/o5-go/deployer/v1/deployer_pb"
 	"github.com/pentops/o5-go/deployer/v1/deployer_tpb"
-	"github.com/pentops/outbox.pg.go/outbox"
+	"google.golang.org/protobuf/encoding/protojson"
 	"google.golang.org/protobuf/proto"
 )
 
@@ -18,182 +19,155 @@ import (
 // same process. Used when running as a standalone tool, e.g. when
 // bootstrapping a new environment.
 type LocalEventLoop struct {
-	messages  chan proto.Message
-	handlers  map[string]func(context.Context, proto.Message) error
 	validator *protovalidate.Validator
-	wg        *sync.WaitGroup
+	storage   *LocalStateStore
+	awsRunner *awsinfra.LocalRunner
 }
 
-func NewLocalEventLoop() *LocalEventLoop {
+func NewLocalEventLoop(awsRunner *awsinfra.LocalRunner) *LocalEventLoop {
 	validator, err := protovalidate.New()
 	if err != nil {
 		panic(err)
 	}
 	return &LocalEventLoop{
+		awsRunner: awsRunner,
 		validator: validator,
-		messages:  make(chan proto.Message),
-		handlers:  map[string]func(context.Context, proto.Message) error{},
-		wg:        &sync.WaitGroup{},
 	}
 }
 
-// PublishEvent adds the event into the processing queue, blocking the Wait.
-func (lel *LocalEventLoop) PublishEvent(ctx context.Context, msg outbox.OutboxMessage) error {
-	if err := lel.validator.Validate(msg); err != nil {
+func (lel *LocalEventLoop) Run(ctx context.Context, trigger *deployer_tpb.TriggerDeploymentMessage) error {
+	if err := lel.validator.Validate(trigger); err != nil {
 		return err
 	}
-	log.WithField(ctx, "inputMessage", msg.ProtoReflect().Descriptor().FullName()).Debug("PublishEvent")
 
-	return lel.handleMessage(ctx, msg)
-}
-
-// Wait runs the event loop. It exits when all handlers have completed, or any
-// handler returns an error. Messages which are queued in PublishEvent prior to
-// a handle exiting extends the lifetime of the event loop.
-func (lel *LocalEventLoop) Wait(ctx context.Context) error {
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-	chErr := make(chan error)
-
-	go func() {
-		for {
-			select {
-			case <-ctx.Done():
-				chErr <- nil
-			case msg := <-lel.messages:
-				go func(msg proto.Message) {
-					err := lel.handleMessage(ctx, msg)
-					if err != nil {
-						chErr <- err
-					}
-				}(msg)
-			}
-		}
-	}()
-
-	err := <-chErr
+	outerEvent, err := deployer.TranslateTrigger(trigger)
 	if err != nil {
-		log.WithError(ctx, err).Error("Event Loop Exit")
-	} else {
-		log.Info(ctx, "Event Loop Exit")
-	}
-	cancel()
-	return err
-}
-
-func (lel *LocalEventLoop) handleMessage(ctx context.Context, msg proto.Message) error {
-	msgKey := string(msg.ProtoReflect().Descriptor().FullName())
-	handler, ok := lel.handlers[msgKey]
-	if !ok {
-		return fmt.Errorf("no handler for message %s", msgKey)
-	}
-
-	handlerContext := log.WithFields(ctx, map[string]interface{}{
-		"inputMessage": msg.ProtoReflect().Descriptor().FullName(),
-	})
-
-	if err := handler(handlerContext, msg); err != nil {
-		if !errors.Is(err, context.Canceled) {
-			log.WithError(handlerContext, err).Error("Event Loop Handler Error")
-		}
 		return err
 	}
-	return nil
-}
 
-func (lel *LocalEventLoop) RegisterHandler(fullName string, handler func(context.Context, proto.Message) error) {
-	lel.handlers[fullName] = handler
-}
+	tx := lel.storage
 
-func wrapHandler[T proto.Message](handler func(context.Context, T) error) (string, func(context.Context, proto.Message) error) {
-	msg := *new(T) // wrapHandler exists for this magic line.
-	fullName := string(msg.ProtoReflect().Descriptor().FullName())
-
-	return fullName, func(ctx context.Context, msg proto.Message) error {
-		ctx = log.WithFields(ctx, map[string]interface{}{
-			"inputMessage": msg.ProtoReflect().Descriptor().FullName(),
-		})
-		typed, ok := msg.(T)
-		if !ok {
-			return fmt.Errorf("unexpected message type: %T", msg)
-		}
-		return handler(ctx, typed)
+	environment, err := tx.GetEnvironment(ctx, trigger.Spec.EnvironmentName)
+	if err != nil {
+		return err
 	}
-}
 
-// RegisterDeployerHandlers takes the proto interface of deployer, and registers
-// the known callbacks to the event loop. This mirrors the standard 'register'
-// pattern for gRPC workers.
-func RegisterDeployerHandlers(eventGroup *LocalEventLoop, deployer deployer_tpb.DeployerTopicServer) error {
+	deployerResolver, err := deployer.BuildParameterResolver(ctx, environment)
+	if err != nil {
+		return err
+	}
 
-	eventGroup.RegisterHandler(wrapHandler(
-		func(ctx context.Context, msg *deployer_tpb.StackStatusChangedMessage) error {
-			_, err := deployer.StackStatusChanged(ctx, msg)
+	eventQueue := make([]*deployer_pb.DeploymentEvent, 0)
+	eventQueue = append(eventQueue, outerEvent)
+
+	for len(eventQueue) > 0 {
+		innerEvent := eventQueue[0]
+		eventQueue = eventQueue[1:]
+
+		deployment, err := tx.GetDeployment(ctx, outerEvent.DeploymentId)
+		if errors.Is(err, deployer.DeploymentNotFoundError) {
+			deployment = &deployer_pb.DeploymentState{
+				DeploymentId: outerEvent.DeploymentId,
+			}
+		} else if err != nil {
 			return err
-		}))
+		}
 
-	eventGroup.RegisterHandler(wrapHandler(
-		func(ctx context.Context, msg *deployer_tpb.MigrationStatusChangedMessage) error {
-			_, err := deployer.MigrationStatusChanged(ctx, msg)
-			return err
-		}))
+		baton := &deployer.TransitionData{
+			ParameterResolver: deployerResolver,
+		}
 
-	eventGroup.RegisterHandler(wrapHandler(
-		func(ctx context.Context, msg *deployer_tpb.TriggerDeploymentMessage) error {
-			_, err := deployer.TriggerDeployment(ctx, msg)
+		typeKey, _ := outerEvent.Event.TypeKey()
+		stateBefore := deployment.Status.ShortString()
+
+		ctx = log.WithFields(ctx, map[string]interface{}{
+			"deploymentId": innerEvent.DeploymentId,
+			"eventType":    typeKey,
+			"transition":   fmt.Sprintf("%s -> ? : %s", stateBefore, typeKey),
+		})
+		log.WithField(ctx, "event", protojson.Format(innerEvent.Event)).Debug("Begin Deployment Event")
+
+		transition, err := deployer.FindTransition(ctx, deployment, innerEvent)
+		if err != nil {
 			return err
-		}))
+		}
+		if err := transition.RunTransition(ctx, baton, deployment, innerEvent); err != nil {
+			return err
+		}
+
+		if err := tx.StoreDeploymentEvent(ctx, deployment, innerEvent); err != nil {
+			return err
+		}
+
+		// Each transiton will produce either one chain event, or one side
+		// effect, until the deployment is terminal.
+		// Each of the side effect handlers will take an action then return a
+		// single status result which triggers the next message.
+		// This is not true of state machines always, but in this case the logic
+		// of the transitions and effects is constrainted to make it possible
+		// to run locally without tracking a complex event loop (and figuring
+		// out when to exit)
+
+		if len(baton.ChainEvents) > 0 && len(baton.SideEffects) > 0 {
+			return fmt.Errorf("cannot have both side effects and chained events in local run mode")
+		}
+
+		if len(baton.ChainEvents) > 0 {
+			eventQueue = append(eventQueue, baton.ChainEvents...)
+			continue
+		}
+
+		for _, sideEffect := range baton.SideEffects {
+			result, err := lel.handleSideEffect(ctx, sideEffect)
+			if err != nil {
+				return err
+			}
+			mapped, err := mapSideEffectResult(result)
+			if err != nil {
+				return err
+			}
+			eventQueue = append(eventQueue, mapped)
+		}
+
+	}
 
 	return nil
+}
+
+func mapSideEffectResult(result proto.Message) (*deployer_pb.DeploymentEvent, error) {
+
+	switch result := result.(type) {
+	case *deployer_tpb.StackStatusChangedMessage:
+		return deployer.TranslateStackStatusChanged(result)
+
+	case *deployer_tpb.MigrationStatusChangedMessage:
+		return deployer.TranslateMigrationStatusChanged(result)
+
+	default:
+		return nil, fmt.Errorf("unknown side effect result type: %T", result)
+	}
 
 }
 
-// RegisterLocalHandlers takes the specific LocalStateStore implementation, as
-// the LSS uses polling loops after triggering AWS actions to emulate
-// subscription to AWS events over sns/sqs.
-func RegisterLocalHandlers(eventGroup *LocalEventLoop, awsRunner *awsinfra.AWSRunner) error {
+func (lel *LocalEventLoop) handleSideEffect(ctx context.Context, msg proto.Message) (proto.Message, error) {
+	log.WithField(ctx, "inputMessage", msg.ProtoReflect().Descriptor().FullName()).Debug("Side Effect")
 
-	eventGroup.RegisterHandler(wrapHandler(
-		func(ctx context.Context, msg *deployer_tpb.UpdateStackMessage) error {
-			_, err := awsRunner.UpdateStack(ctx, msg)
-			return err
-		}))
+	switch msg := msg.(type) {
+	case *deployer_tpb.UpdateStackMessage:
+		return lel.awsRunner.UpdateStack(ctx, msg)
 
-	eventGroup.RegisterHandler(wrapHandler(
-		func(ctx context.Context, msg *deployer_tpb.CreateNewStackMessage) error {
-			_, err := awsRunner.CreateNewStack(ctx, msg)
-			return err
-		}))
+	case *deployer_tpb.CreateNewStackMessage:
+		return lel.awsRunner.CreateNewStack(ctx, msg)
 
-	eventGroup.RegisterHandler(wrapHandler(
-		func(ctx context.Context, msg *deployer_tpb.DeleteStackMessage) error {
-			_, err := awsRunner.DeleteStack(ctx, msg)
-			return err
-		}))
+	case *deployer_tpb.ScaleStackMessage:
+		return lel.awsRunner.ScaleStack(ctx, msg)
 
-	eventGroup.RegisterHandler(wrapHandler(
-		func(ctx context.Context, msg *deployer_tpb.ScaleStackMessage) error {
-			_, err := awsRunner.ScaleStack(ctx, msg)
-			return err
-		}))
+	case *deployer_tpb.StabalizeStackMessage:
+		return lel.awsRunner.StabalizeStack(ctx, msg)
 
-	eventGroup.RegisterHandler(wrapHandler(
-		func(ctx context.Context, msg *deployer_tpb.CancelStackUpdateMessage) error {
-			_, err := awsRunner.CancelStackUpdate(ctx, msg)
-			return err
-		}))
+	case *deployer_tpb.RunDatabaseMigrationMessage:
+		return lel.awsRunner.RunDatabaseMigration(ctx, msg)
+	}
 
-	eventGroup.RegisterHandler(wrapHandler(
-		func(ctx context.Context, msg *deployer_tpb.StabalizeStackMessage) error {
-			_, err := awsRunner.StabalizeStack(ctx, msg)
-			return err
-		}))
-
-	eventGroup.RegisterHandler(wrapHandler(
-		func(ctx context.Context, msg *deployer_tpb.RunDatabaseMigrationMessage) error {
-			_, err := awsRunner.RunDatabaseMigration(ctx, msg)
-			return err
-		}))
-
-	return nil
+	return nil, fmt.Errorf("unknown side effect message type: %T", msg)
 }
