+++ localrun/statestore.go
@@ -8,12 +8,7 @@ import (
 	"github.com/pentops/o5-deploy-aws/deployer"
 	"github.com/pentops/o5-go/deployer/v1/deployer_pb"
 	"github.com/pentops/o5-go/environment/v1/environment_pb"
-	"github.com/pentops/outbox.pg.go/outbox"
 )
-
-type Publisher interface {
-	PublishEvent(ctx context.Context, msg outbox.OutboxMessage) error
-}
 
 // LocalStateStore wires back the events to the deployer, rather than relying on
 // an event bus and database
@@ -23,8 +18,6 @@ type LocalStateStore struct {
 	deployments  map[string]*deployer_pb.DeploymentState
 	environments map[string]*environment_pb.Environment
 
-	eventLoop Publisher
-
 	validator *protovalidate.Validator
 
 	waitChan chan error
@@ -32,7 +25,7 @@ type LocalStateStore struct {
 	StoreCallback func(ctx context.Context, state *deployer_pb.DeploymentState, event *deployer_pb.DeploymentEvent) error
 }
 
-func NewLocalStateStore(eventLoop Publisher) *LocalStateStore {
+func NewLocalStateStore() *LocalStateStore {
 	validator, err := protovalidate.New()
 	if err != nil {
 		panic(err)
@@ -42,20 +35,10 @@ func NewLocalStateStore(eventLoop Publisher) *LocalStateStore {
 		stackMap:     map[string]string{},
 		environments: map[string]*environment_pb.Environment{},
 		deployments:  map[string]*deployer_pb.DeploymentState{},
-		eventLoop:    eventLoop,
 		validator:    validator,
 		waitChan:     make(chan error),
 	}
 	return lss
-}
-
-func (lss *LocalStateStore) Wait(ctx context.Context) error {
-	select {
-	case <-ctx.Done():
-		return nil
-	case err := <-lss.waitChan:
-		return err
-	}
 }
 
 func (lss *LocalStateStore) AddEnvironment(environment *environment_pb.Environment) error {
@@ -67,20 +50,11 @@ func (lss *LocalStateStore) AddEnvironment(environment *environment_pb.Environme
 	return nil
 }
 
-func (lss *LocalStateStore) Transact(ctx context.Context, fn func(context.Context, deployer.TransitionTransaction) error) error {
-	// Local state doesn't need a transaction as it doesn't have a database
-	return fn(ctx, lss)
-}
-
 func (lss *LocalStateStore) GetEnvironment(ctx context.Context, environmentName string) (*environment_pb.Environment, error) {
 	if env, ok := lss.environments[environmentName]; ok {
 		return env, nil
 	}
 	return nil, fmt.Errorf("missing environment %s", environmentName)
-}
-
-func (lss *LocalStateStore) PublishEvent(ctx context.Context, msg outbox.OutboxMessage) error {
-	return lss.eventLoop.PublishEvent(ctx, msg)
 }
 
 func (lss *LocalStateStore) StoreDeploymentEvent(ctx context.Context, state *deployer_pb.DeploymentState, event *deployer_pb.DeploymentEvent) error {
